Monitor Interval type changed from uint16 to uint32 in itm-config.yang
as per RFC 5880 BFD Desired Min TX Interval is a 32 bit value specified in
microseconds .odl-interface.yang has uin32 as datatype for monitor
interval.This code changes will create sync in datatype for tunnel
monitoring interval.

+++ itm/itm-api/src/main/java/org/opendaylight/genius/itm/globals/ITMConstants.java
@@ -22,10 +22,10 @@ public class ITMConstants {
     public static final int LLDP_SERVICE_ID = 0;
     // Tunnel Monitoring
     public static final boolean DEFAULT_MONITOR_ENABLED = true;
-    public static final int DEFAULT_MONITOR_INTERVAL = 10000;
-    public static final int BFD_DEFAULT_MONITOR_INTERVAL = 1000;
-    public static final int MIN_MONITOR_INTERVAL = 1000;
-    public static final int MAX_MONITOR_INTERVAL = 30000;
+    public static final long DEFAULT_MONITOR_INTERVAL = 10000L;
+    public static final long BFD_DEFAULT_MONITOR_INTERVAL = 1000L;
+    public static final long MIN_MONITOR_INTERVAL = 1000L;
+    public static final long MAX_MONITOR_INTERVAL = 30000L;
     public static final String DUMMY_IP_ADDRESS = "0.0.0.0";
     public static final String TUNNEL_TYPE_VXLAN = "VXLAN";
     public static final String TUNNEL_TYPE_GRE = "GRE";
